LayoutTests/fullscreen/full-screen-iframe-allowed-expected.txt

Test for bug 56264: Handle entering full screen security restrictions

To test manually, press Space - the page should enter full screen mode.

EVENT(webkitfullscreenchange)
SUCCEED - entered full screen!
END OF TEST
